Feat: add di module and set file struct
### [Feat]: [DI를 위한 작업과 파일 구조 세팅]


# 1. DI setting
- App Module, App 생성

# 2. 파일 구조 세팅
data, domain, presentation layer로 나눔
