+++ src/BlogApp/app/src/main/java/com/smilegateblog/smliegateblog/di/NetworkModule.kt
@@ -1,0 +1,28 @@
+package com.smilegateblog.smliegateblog.di
+
+import com.smilegateblog.smliegateblog.BuildConfig
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+class NetworkModule {
+
+    @Singleton
+    @Provides
+    fun provideRetrofit(okHttpClient: OkHttpClient) : Retrofit {
+        return Retrofit.Builder().apply {
+            addConverterFactory(GsonConverterFactory.create())
+            baseUrl(BuildConfig.API_BASE_URL)
+        }.build()
+    }
+
+}
