+++ src/BlogApp/app/src/main/java/com/smilegateblog/smliegateblog/presentation/App.kt
@@ -1,0 +1,11 @@
+package com.smilegateblog.smliegateblog.presentation
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application(){
+    override fun onCreate() {
+        super.onCreate()
+    }
+}
