+++ src/BlogApp/app/src/main/java/com/smilegateblog/smliegateblog/util/Resource.kt
@@ -1,0 +1,10 @@
+package com.smilegateblog.smliegateblog.util
+
+sealed class Resource<T>(val data: T? = null, val message: String? = null) {
+    class Success<T>(data: T) : Resource<T>(data)
+    class Error<T>(message: String, data: T? = null) : Resource<T>(data, message)
+    class Loading<T>(data: T? = null) : Resource<T>(data)
+
+    class NetworkConnectionError<T>(message: String = "Network Error", data: T? = null) :
+        Resource<T>(data, message)
+}
